Synthetic Societies/requirements.txt
autogen-agentchat==0.6.4
autogen-core==0.6.4
autogen-ext==0.6.4
openai==1.35.7
httpx==0.26.0
tiktoken>=0.5.0,<0.13.0
numpy>=1.23,<2.3
matplotlib>=3.7,<3.11
huggingface_hub>=0.21.0
python-dotenv>=1.0.0
wget
transformers>=4.40.0
sentencepiece
